code-style: reorganize the repository
While the diff is pretty large, this commit
is just about splitting the base.py and its associated
test file into various scope-restricted files.
The api folder is also introduced.

This will ease the development/reviews
