+++ consul/__init__.py
@@ -1,4 +1,5 @@
 __version__ = "1.3.0"
 
-from consul.base import ACLDisabled, ACLPermissionDenied, Check, ConsulException, NotFound, Timeout
+from consul.check import Check
+from consul.exceptions import ACLDisabled, ACLPermissionDenied, ConsulException, NotFound, Timeout
 from consul.std import Consul
