+++ tests/api/test_txn.py
@@ -1,0 +1,14 @@
+import base64
+
+
+class TestTxn:
+    def test_transaction(self, consul_obj):
+        c, _consul_version = consul_obj
+        value = base64.b64encode(b"1").decode("utf8")
+        d = {"KV": {"Verb": "set", "Key": "asdf", "Value": value}}
+        r = c.txn.put([d])
+        assert r["Errors"] is None
+
+        d = {"KV": {"Verb": "get", "Key": "asdf"}}
+        r = c.txn.put([d])
+        assert r["Results"][0]["KV"]["Value"] == value
